# public/apidoc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APD-API",
        "description": "An API to perform air pressure drop calculation through projects and user account",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://aeraulic.io/api/V1",
            "description": "Apd calculator API server"
        }
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "App user"
                ],
                "summary": "Get user datas",
                "description": "Get the user's datas",
                "operationId": "84e675ece83719d4df8b82e07d6785ee",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/appUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "App user"
                ],
                "summary": "Remove user",
                "description": "Perform a full removing of user's data and all entity related (projects, duct networks and duct sections).",
                "operationId": "574d27c5a9727a0fd17708de1ecdf593",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Remove an app user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User \"email@email.test\" and all projects are deleted."
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/appUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "App user"
                ],
                "summary": "Update user",
                "description": "Update the user's datas.",
                "operationId": "0ed50950c8f15ccf4bd38308345c5ee8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateAppUser"
                },
                "responses": {
                    "200": {
                        "description": "Update an app user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/appUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessableEntity"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "App user"
                ],
                "summary": "authenticated user's datas",
                "description": "Get the authenticated user's data",
                "operationId": "a1f5836cf1aed4fd0f9c45f2a5b5ffaf",
                "responses": {
                    "200": {
                        "description": "Get user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/appUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "App user"
                ],
                "summary": "Register user",
                "description": "Register request for a visitor who wants a account to use the application.",
                "operationId": "eef400eb9b9887436475a1ff5e7b096c",
                "requestBody": {
                    "$ref": "#/components/requestBodies/register"
                },
                "responses": {
                    "200": {
                        "description": "Register user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/appUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register/{id}": {
            "post": {
                "tags": [
                    "App user"
                ],
                "summary": "Confirm register",
                "description": "Confirm register from a user with the email sent on registration.",
                "operationId": "a3a1587c041d2ad0b919688c53673d23",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirm user's registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User correctly activated"
                                        },
                                        "content": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User already registrated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "content": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "field": {
                                                        "type": "string",
                                                        "example": "user"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "User is already active"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "App user"
                ],
                "summary": "Login",
                "description": "Login request to authentifiate user.",
                "operationId": "a79ced35ad4f96452b566c8199972d70",
                "requestBody": {
                    "$ref": "#/components/requestBodies/login"
                },
                "responses": {
                    "200": {
                        "description": "Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "title": "token",
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpYXQiOjE3MDMzNDY2MjQsImV4cCI6MTcwMzQzMzAyNCwicm9sZXMiOlsiUk9MRV9VU0VSIl0sInVzZXJuYW1lIjoidGVzdEBlbWFpbC50ZXN0In0.H-ZOfO9x5uH1VKo33s4Nb1UXcKG5ZIlaayu9OyRdeuMDcvMiEgozizgR9l-63G0uNkDOpOcXE4aeXk7RH0Lwj7c8OZN-KHxkYYcIDIJ-dDM4I72ur2z87EW9K0VhjtUxiCylzlMKcybHG1idixyGeaOTHx0fAH5KLT6KBsFgbboZoIKc9TLjH59eNBiLob0r_fssLdpEDqhWA4AWgN7NHRmIg7EXbEu54QEXPs3n7_vvN_QKkR6YltoyQDQ4Lj5dmWpYtcZOLNi9fLq91VOjFB0gTTuyS-SviOxuXtoJzgPbDaf6Q7BUYlBgDRk2Alegru9IQggHeMjBm0IG3Sk0Cg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "title": "code http",
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "appUser": {
                "title": "app user",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "user id",
                        "type": "integer",
                        "example": 10
                    },
                    "email": {
                        "title": "email",
                        "description": "User's email",
                        "type": "string",
                        "format": "email",
                        "example": "email@email.test"
                    },
                    "password": {
                        "title": "password",
                        "description": "User's password - must have minimum 8 characters, 1 uppercase letter, 1 lowercase letter, 1 digit and 1 special character",
                        "type": "string",
                        "pattern": "/^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%&^*-]).{8,}$/",
                        "example": "$2y$10$./m7/hhKacEd85Vu5tZeHOd9r4sUYF375GHRqEhaen/sfoixxYtRi"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "User's last name",
                        "type": "string",
                        "example": "Doe"
                    },
                    "firstname": {
                        "title": "firstname",
                        "description": "User's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "company": {
                        "title": "company",
                        "description": "User's company - optional",
                        "type": "string",
                        "example": "Johns&Cie",
                        "nullable": true
                    },
                    "role": {
                        "title": "role",
                        "description": "User's role",
                        "type": "string",
                        "example": "appUser"
                    },
                    "isEnable": {
                        "title": "isEnable",
                        "description": "Is user enable",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Ressource's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}